got the translocation screen working properly
